# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Event": {
        "properties": {
          "app_user_id": {
            "type": "string"
          },
          "entitlement_ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "environment": {
            "type": "string"
          },
          "expiration_at_ms": {
            "format": "int64",
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "original_app_user_id": {
            "type": "string"
          },
          "product_id": {
            "type": "string"
          },
          "purchased_at_ms": {
            "format": "int64",
            "type": "integer"
          },
          "store": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "MediaRatingSummaryResponse": {
        "properties": {
          "averageRating": {
            "format": "double",
            "type": "number"
          },
          "mediaId": {
            "format": "int64",
            "type": "integer"
          },
          "ratingCount": {
            "format": "int64",
            "type": "integer"
          },
          "starBreakdown": {
            "additionalProperties": {
              "format": "int64",
              "type": "integer"
            },
            "type": "object"
          }
        },
        "type": "object"
      },
      "MediaReviewResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "mediaId": {
            "format": "int64",
            "type": "integer"
          },
          "rating": {
            "format": "int32",
            "type": "integer"
          },
          "reviewText": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "userId": {
            "format": "int64",
            "type": "integer"
          },
          "userInitial": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "MediaReviewSaveRequest": {
        "properties": {
          "rating": {
            "format": "int32",
            "maximum": 5,
            "minimum": 1,
            "type": "integer"
          },
          "reviewText": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          }
        },
        "type": "object"
      },
      "PageMediaReviewResponse": {
        "properties": {
          "content": {
            "items": {
              "$ref": "#/components/schemas/MediaReviewResponse"
            },
            "type": "array"
          },
          "empty": {
            "type": "boolean"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "number": {
            "format": "int32",
            "type": "integer"
          },
          "numberOfElements": {
            "format": "int32",
            "type": "integer"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "totalElements": {
            "format": "int64",
            "type": "integer"
          },
          "totalPages": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "Pageable": {
        "properties": {
          "page": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "minimum": 1,
            "type": "integer"
          },
          "sort": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "PageableObject": {
        "properties": {
          "offset": {
            "format": "int64",
            "type": "integer"
          },
          "pageNumber": {
            "format": "int32",
            "type": "integer"
          },
          "pageSize": {
            "format": "int32",
            "type": "integer"
          },
          "paged": {
            "type": "boolean"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "unpaged": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "ProfileDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "notificationPreferences": {
            "type": "string"
          },
          "profileId": {
            "format": "int64",
            "type": "integer"
          },
          "themePreference": {
            "type": "string"
          },
          "twoFactorEnabled": {
            "type": "boolean"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "userId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "RevenueCatWebhookRequest": {
        "properties": {
          "api_version": {
            "type": "string"
          },
          "event": {
            "$ref": "#/components/schemas/Event"
          }
        },
        "type": "object"
      },
...
      "SortObject": {
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Keycloak-issued access token. Two realms are accepted, the mobile realm and the admin realm; the token's issuer selects which. Endpoints under /admin additionally require the ADMIN role.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Yogida Team",
      "url": "https://github.com/yogida"
    },
    "description": "REST API for the Yogida meditation platform. \nProvides presigned Cloudflare R2 streaming URLs and a catalog of available meditation media grouped by bucket.",
    "title": "Meditation Service API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/media/{mediaId}/rating-summary": {
      "get": {
        "description": "Returns the average rating, total rating count, and per-star breakdown (keys 1–5) for a media item.\nAll keys 1–5 are present in `starBreakdown` (missing stars have a count of 0).\n",
        "operationId": "getMediaRatingSummary",
        "parameters": [
          {
            "description": "Media ID",
            "in": "path",
            "name": "mediaId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaRatingSummaryResponse"
                }
              }
            },
            "description": "Rating summary returned"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaRatingSummaryResponse"
                }
              }
            },
            "description": "Media not found"
          }
        },
        "summary": "Get rating summary",
        "tags": [
          "Media — Ratings & Reviews"
        ]
      }
    },
    "/media/{mediaId}/review": {
      "put": {
        "description": "Creates or updates the review/rating row for the given user and media item.\n`rating` updates the star rating (1–5) on every call.\n`reviewText` is written on first save only; subsequent saves do not change it.\n",
        "operationId": "saveMediaReview",
        "parameters": [
          {
            "description": "Media ID",
            "in": "path",
            "name": "mediaId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaReviewSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaReviewResponse"
                }
              }
            },
            "description": "Review saved"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaReviewResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaReviewResponse"
                }
              }
            },
            "description": "Media or user not found"
          }
        },
        "summary": "Save review and/or rating",
        "tags": [
          "Media — Ratings & Reviews"
        ]
      }
    },
    "/media/{mediaId}/reviews": {
      "get": {
        "description": "Returns a paginated list of reviews for a media item.\nEach entry may have a null `rating` (text-only review) or null `reviewText` (rating-only entry).\n`userName` is derived from the user email prefix; `userInitial` is the first uppercase character.\nSupported sort fields: `createdAt` (default, desc), `rating` (asc or desc).\nExample: GET /media/1/reviews?page=0&size=10&sort=rating,desc\n",
        "operationId": "listMediaReviews",
        "parameters": [
          {
            "description": "Media ID",
            "in": "path",
            "name": "mediaId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageable",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageMediaReviewResponse"
                }
              }
            },
            "description": "Reviews page returned"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageMediaReviewResponse"
                }
              }
            },
            "description": "Unsupported sort field"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageMediaReviewResponse"
                }
              }
            },
            "description": "Media not found"
          }
        },
        "summary": "List reviews",
        "tags": [
          "Media — Ratings & Reviews"
        ]
      }
    },
    "/profiles": {
      "get": {
        "description": "Returns a list of all user profiles.",
        "operationId": "listProfiles",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProfileDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Profiles retrieved successfully"
          }
        },
        "summary": "Get all profiles",
        "tags": [
          "Profiles"
        ]
      },
      "post": {
        "description": "Creates a new user profile.",
        "operationId": "createProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            },
            "description": "Profile created successfully"
          }
        },
        "summary": "Create a new profile",
        "tags": [
          "Profiles"
        ]
      }
    },
    "/profiles/{id}": {
      "delete": {
        "description": "Removes a profile by ID.",
        "operationId": "deleteProfile",
        "parameters": [
          {
            "description": "Profile ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Profile deleted successfully"
          },
          "404": {
            "description": "Profile not found"
          }
        },
        "summary": "Delete a profile",
        "tags": [
          "Profiles"
        ]
      },
      "get": {
        "description": "Returns a single profile by its ID.",
        "operationId": "getProfileById",
        "parameters": [
          {
            "description": "Profile ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            },
            "description": "Profile found"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            },
            "description": "Profile not found"
          }
        },
        "summary": "Get profile by ID",
        "tags": [
          "Profiles"
        ]
      },
      "put": {
        "description": "Updates profile details by ID.",
        "operationId": "updateProfile",
        "parameters": [
          {
            "description": "Profile ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            },
            "description": "Profile updated successfully"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            },
            "description": "Profile not found"
          }
        },
        "summary": "Update an existing profile",
        "tags": [
          "Profiles"
        ]
      }
    },
...
    "/webhooks/revenuecat": {
      "post": {
        "description": "Authenticated by a shared secret in the Authorization header. Idempotent per event id.",
        "operationId": "handleRevenueCatWebhook",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevenueCatWebhookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Event accepted"
          },
          "401": {
            "description": "Missing or invalid webhook authorization"
          }
        },
        "summary": "Process a RevenueCat webhook event",
        "tags": [
          "RevenueCat Webhooks"
        ]
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Current environment",
      "url": "/api"
    }
  ],
...
}